chatbot/requirements.txt
langchain==0.1.0
langchain-community==0.0.13
langchain-google-genai==0.0.5
faiss-cpu==1.7.4
python-dotenv==1.0.0
google-generativeai==0.3.1
pandas==2.1.1
numpy==1.26.0
python-docx==1.0.1
pyPDF2==3.0.1